+++ client/web/src/integration/batches.test.ts
@@ -469,11 +469,11 @@ describe('Batches', () => {
                 await driver.page.waitForSelector('.test-file-diff-node')
 
                 // Switch to view burndown chart.
-                await driver.page.click('.test-batches-chart-tab')
+                await driver.page.click('[data-reach-tab-list] button:nth-child(2)')
                 await driver.page.waitForSelector('.test-batches-chart')
 
                 // Switch to view spec file.
-                await driver.page.click('.test-batches-spec-tab')
+                await driver.page.click('[data-reach-tab-list] button:nth-child(3)')
                 await driver.page.waitForSelector('.test-batches-spec')
 
                 // Go to close page via button.
@@ -571,6 +571,7 @@ describe('Batches', () => {
                                       },
                             supersedingBatchSpec: null,
                             viewerCanAdminister: true,
+                            originalInput: 'name: awesome-batch-change\ndescription: somestring',
                             applyPreview: {
                                 stats: {
                                     close: 10,
@@ -587,6 +588,7 @@ describe('Batches', () => {
                                     modified: 10,
                                     removed: 3,
                                 },
+                                totalCount: 10,
                             },
                             viewerBatchChangesCodeHosts: {
                                 totalCount: 0,