+++ client/web/src/search/input/SearchContextMenu.tsx
@@ -21,29 +21,9 @@ import { useObservable } from '@sourcegraph/shared/src/util/useObservable'
 import { SearchContextInputProps } from '..'
 import { SearchContextFields } from '../../graphql-operations'
 
-const HighlightedSearchTerm: React.FunctionComponent<{ text: string; searchFilter: string }> = ({
-    text,
-    searchFilter,
-}) => {
-    if (searchFilter.length > 0) {
-        const index = text.toLowerCase().indexOf(searchFilter.toLowerCase())
-        if (index > -1) {
-            const before = text.slice(0, index)
-            const highlighted = text.slice(index, index + searchFilter.length)
-            const after = text.slice(index + searchFilter.length)
-            return (
-                <>
-                    {before}
-                    <strong className="search-context-menu__item--highlighted">{highlighted}</strong>
-                    {after}
-                </>
-            )
-        }
-    }
-    return <>{text}</>
-}
+import { HighlightedSearchContextSpec } from './HighlightedSearchContextSpec'
 
-const SearchContextMenuItem: React.FunctionComponent<{
+export const SearchContextMenuItem: React.FunctionComponent<{
     spec: string
     description: string
     selected: boolean
@@ -60,7 +40,7 @@ const SearchContextMenuItem: React.FunctionComponent<{
             onClick={setContext}
         >
             <small className="search-context-menu__item-name font-weight-medium" title={spec}>
-                <HighlightedSearchTerm text={spec} searchFilter={searchFilter} />
+                <HighlightedSearchContextSpec spec={spec} searchFilter={searchFilter} />
             </small>{' '}
             <small className="search-context-menu__item-description" title={description}>
                 {description}
@@ -90,7 +70,7 @@ interface NextPageUpdate {
 
 type LoadingState = 'LOADING' | 'LOADING_NEXT_PAGE' | 'DONE' | 'ERROR'
 
-const searchContextsPerPageToLoad = 10
+const searchContextsPerPageToLoad = 15
 
 const getFirstMenuItem = (): HTMLButtonElement | null =>
     document.querySelector('.search-context-menu__item:first-child')
@@ -312,7 +292,7 @@ export const SearchContextMenu: React.FunctionComponent<SearchContextMenuProps> 
                         <small>No contexts found</small>
                     </DropdownItem>
                 )}
-                <div ref={infiniteScrollTrigger} />
+                <div ref={infiniteScrollTrigger} className="search-context-menu__infinite-scroll-trigger" />
             </div>
             <div className="search-context-menu__footer">
                 <button