+++ internal/conf/parse.go
@@ -57,6 +57,7 @@ var requireRestart = []string{
 	"auth.providers",
 	"externalURL",
 	"update.channel",
+	"insights.query.worker.concurrency",
 }
 
 // NeedRestartToApply determines if a restart is needed to apply the changes